Add unit test to improve branch coverage
Coverage analysis indicated that lines 123-126 in lib/Test/Vars.pm (an
'else' branch within _check_vars()) were not being exercised by the test
suite.  Adding a test in which the argument to vars_ok() appears to be a
package name (e.g., 'Test::Vars') exercise this branch and improves
statement and branch coverage.