Add unit test to improve branch coverage
Coverage analysis indicated that the implicit 'else' branch of the 'if'
condition at lines 132 in lib/Test/Vars.pm was not being exercised by
the test suite.  Adding a test in which additional arguments are
provided to vars_ok(), the additions being a key-value pair designed to
exercise the 'else' condition, thereby improving branch coverage.