Remove dependency on 'subdirs-test_dynamic'
It didn't exist in EU::MM until approx. version 7.12, committed to
repository on Apr 29 2016.  So if EU::MM is older than that (as it may
very well be in older perls whose EU::MMs have not been upgraded),
Makefile.PL would gag.  It turns out not to have been needed.